Make func to update io more conformative to hudhook
