+++ src/hooks/common.rs
@@ -11,6 +11,7 @@ use windows::Win32::Foundation::{
     CloseHandle, BOOL, HANDLE, HINSTANCE, HWND, INVALID_HANDLE_VALUE, LPARAM, LRESULT, POINT, RECT,
     WPARAM,
 };
+use windows::Win32::Graphics::Gdi::ScreenToClient;
 use windows::Win32::System::Diagnostics::ToolHelp::{
     CreateToolhelp32Snapshot, Thread32First, Thread32Next, TH32CS_SNAPTHREAD, THREADENTRY32,
 };
@@ -72,6 +73,60 @@ pub(crate) trait ImguiWindowsEventHandler {
         io[Key::X] = VK_X.0 as _;
         io[Key::Y] = VK_Y.0 as _;
         io[Key::Z] = VK_Z.0 as _;
+    }
+
+    unsafe fn update_io(
+        &mut self,
+        render_loop: &mut Box<dyn ImguiRenderLoop + Send + Sync>,
+        game_hwnd: HWND,
+        window_rect: RECT,
+    ) {
+        let mut io = ImguiWindowsEventHandler::io_mut(self);
+
+        for i in 0..256 {
+            io.keys_down[i] = is_key_down(i);
+        }
+
+        for i in 0..5 {
+            io.mouse_down[i] = is_mouse_button_down(i);
+        }
+
+        let char = INPUT_CHARACTER.swap(0, Ordering::SeqCst);
+
+        if char != 0 {
+            io.add_input_character(char as char);
+        }
+
+        io.mouse_wheel += MOUSE_WHEEL_DELTA.swap(0, Ordering::SeqCst) as f32;
+        io.mouse_wheel_h += MOUSE_WHEEL_DELTA_H.swap(0, Ordering::SeqCst) as f32;
+
+        if render_loop.should_block_messages(&io) {
+            if !io.mouse_draw_cursor {
+                io.mouse_draw_cursor = true;
+                GAME_MOUSE_BLOCKED.store(true, Ordering::SeqCst);
+            }
+        } else {
+            if io.mouse_draw_cursor {
+                io.mouse_draw_cursor = false;
+                GAME_MOUSE_BLOCKED.store(false, Ordering::SeqCst);
+            }
+        }
+
+        io.display_size = [
+            (window_rect.right - window_rect.left) as f32,
+            (window_rect.bottom - window_rect.top) as f32,
+        ];
+        let mut pos = *LAST_CURSOR_POS.get().unwrap().lock();
+
+        let active_window = GetForegroundWindow();
+        if !HANDLE(active_window.0).is_invalid()
+            && (active_window == game_hwnd || IsChild(active_window, game_hwnd).as_bool())
+        {
+            ScreenToClient(active_window, &mut pos as *mut _);
+
+            io.mouse_pos[0] = pos.x as f32;
+            io.mouse_pos[1] = pos.y as f32;
+        }
     }
 }
 
@@ -193,40 +248,6 @@ pub unsafe fn is_mouse_button_down(button: usize) -> bool {
         return is_key_down(VK_LBUTTON.0 as usize + button);
     } else {
         is_key_down(VK_LBUTTON.0 as usize + button + 1)
-    }
-}
-
-pub unsafe fn update_imgui_io(
-    io: &mut Io,
-    render_loop: &mut Box<dyn ImguiRenderLoop + Send + Sync>,
-) {
-    for i in 0..256 {
-        io.keys_down[i] = is_key_down(i);
-    }
-
-    for i in 0..5 {
-        io.mouse_down[i] = is_mouse_button_down(i);
-    }
-
-    let char = INPUT_CHARACTER.swap(0, Ordering::SeqCst);
-
-    if char != 0 {
-        io.add_input_character(char as char);
-    }
-
-    io.mouse_wheel += MOUSE_WHEEL_DELTA.swap(0, Ordering::SeqCst) as f32;
-    io.mouse_wheel_h += MOUSE_WHEEL_DELTA_H.swap(0, Ordering::SeqCst) as f32;
-
-    if render_loop.should_block_messages(&io) {
-        if !io.mouse_draw_cursor {
-            io.mouse_draw_cursor = true;
-            GAME_MOUSE_BLOCKED.store(true, Ordering::SeqCst);
-        }
-    } else {
-        if io.mouse_draw_cursor {
-            io.mouse_draw_cursor = false;
-            GAME_MOUSE_BLOCKED.store(false, Ordering::SeqCst);
-        }
     }
 }
