+++ src/hooks/opengl3.rs
@@ -6,12 +6,12 @@ use log::{debug, trace};
 use once_cell::sync::OnceCell;
 use parking_lot::Mutex;
 use windows::core::PCSTR;
-use windows::Win32::Foundation::{GetLastError, BOOL, HANDLE, HWND, POINT, RECT};
-use windows::Win32::Graphics::Gdi::{ScreenToClient, WindowFromDC, HDC};
+use windows::Win32::Foundation::{GetLastError, BOOL, HWND, POINT, RECT};
+use windows::Win32::Graphics::Gdi::{WindowFromDC, HDC};
 use windows::Win32::System::LibraryLoader::{GetModuleHandleA, GetProcAddress};
-use windows::Win32::UI::WindowsAndMessaging::{GetClientRect, GetForegroundWindow, IsChild};
+use windows::Win32::UI::WindowsAndMessaging::GetClientRect;
 
-use super::common::{self, update_imgui_io, KEYS, LAST_CURSOR_POS};
+use super::common::{self, KEYS, LAST_CURSOR_POS};
 use crate::hooks::common::ImguiWindowsEventHandler;
 use crate::hooks::{Hooks, ImguiRenderLoop, ImguiRenderLoopFlags};
 use crate::mh::{MhHook, MhHooks};
@@ -106,22 +106,7 @@ impl ImguiRenderer {
         let render_loop = IMGUI_RENDER_LOOP.get_mut().unwrap();
 
         if let Some(rect) = get_client_rect(&self.game_hwnd) {
-            let mut io = self.ctx.io_mut();
-            io.display_size = [(rect.right - rect.left) as f32, (rect.bottom - rect.top) as f32];
-            let mut pos = *LAST_CURSOR_POS.get().unwrap().lock();
-
-            update_imgui_io(io, render_loop);
-
-            let active_window = GetForegroundWindow();
-            if !HANDLE(active_window.0).is_invalid()
-                && (active_window == self.game_hwnd
-                    || IsChild(active_window, self.game_hwnd).as_bool())
-            {
-                ScreenToClient(active_window, &mut pos as *mut _);
-
-                io.mouse_pos[0] = pos.x as f32;
-                io.mouse_pos[1] = pos.y as f32;
-            }
+            ImguiWindowsEventHandler::update_io(self, render_loop, self.game_hwnd, rect);
         } else {
             trace!("GetWindowRect error: {:x}", GetLastError().0);
         }
